fix : refine not found handling

+++ src/controllers/banner.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interface";
 import response from "../utils/response";
 import BannerModel, { bannerDAO, TypeBanner } from "../models/banner.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 
 export default {
@@ -55,6 +55,11 @@ export default {
   async findOne (req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed find one a banner");
+      }
+
       const result = await BannerModel.findById(id);
 
       if (!result) {
@@ -71,6 +76,11 @@ export default {
     try {
       const { id } = req.params;
       await bannerDAO.validate(req.body);
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed update a banner");
+      }
+
       const result = await BannerModel.findByIdAndUpdate(id, req.body, { new: true });
       response.success(res, result, "success update a banner");
     } catch (error) {
@@ -81,6 +91,11 @@ export default {
   async remove (req: IReqUser, res: Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed remove a banner");
+      }
+
       const result = await BannerModel.findByIdAndDelete(id, { new: true });
       response.success(res, result, "success remove a banner");
     } catch (error) {
