+++ src/controllers/category.controller.ts
@@ -2,6 +2,7 @@ import { Response } from "express";
 import { IReqUser, IPaginationQuery } from "../utils/interface";
 import CategoryModel, { categoryDAO } from "../models/category.model";
 import response from "../utils/response";
+import { isValidObjectId } from "mongoose";
 
 export default {
   async create (req : IReqUser, res : Response) {
@@ -51,6 +52,9 @@ export default {
   async findOne (req : IReqUser, res : Response) {
     try {
       const { id } = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed find one a category");
+      }
       const result = await CategoryModel.findById(id);
       if (!result) {
         return response.notFound(res, "failed find one a category");
@@ -64,6 +68,11 @@ export default {
   async update (req : IReqUser, res : Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed update a category");
+      }
+
       const result = await CategoryModel.findByIdAndUpdate(id, req.body, { new : true });
       response.success(res, result, "success update a category");
     } catch (error) {
@@ -74,6 +83,11 @@ export default {
   async remove (req : IReqUser, res : Response) {
     try {
       const { id } = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed remove a category");
+      }
+
       const result = await CategoryModel.findByIdAndDelete(id, {new : true});
       response.success(res, result, "success remove a category");
     } catch (error) {
