+++ src/controllers/event.controller.ts
@@ -2,7 +2,7 @@ import { Response } from "express";
 import { IPaginationQuery, IReqUser } from "../utils/interface";
 import response from "../utils/response";
 import EventModel, { eventDAO, TEvent } from "../models/event.model";
-import { FilterQuery } from "mongoose";
+import { FilterQuery, isValidObjectId } from "mongoose";
 
 export default {
   async create(req: IReqUser, res: Response) {
@@ -53,6 +53,9 @@ export default {
   async findOne(req: IReqUser, res: Response) {
     try {
       const {id} = req.params;
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed find one event");
+      }
       const result = await EventModel.findById(id);
       if (!result) {
         return response.notFound(res, "failed find one event");
@@ -66,6 +69,11 @@ export default {
   async update(req: IReqUser, res: Response) {
     try {
       const {id} = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed update one event");
+      }
+
       const result = await EventModel.findByIdAndUpdate(id, req.body, {new: true});
       response.success(res, result, "success update one event");
     } catch (error) {
@@ -76,6 +84,11 @@ export default {
   async remove(req: IReqUser, res: Response) {
     try {
       const {id} = req.params;
+
+      if (!isValidObjectId(id)) {
+        return response.notFound(res, "failed remove one event");
+      }
+
       const result = await EventModel.findByIdAndDelete(id, {new : true});
       response.success(res, result, "success remove one event");
     } catch (error) {
